Added maps for tiles for better accessibility
The actor and staticElement field of level are now maps from identifiers to the
actual data. This gives a level the ability to identify a tile by its id.

I also added a coordinate to list of identifier map. This allows to get all
tiles that locates at a specific coordinate.