added toString, which renders a level to a string
The function has to know the boundaries of the level, so I added a field
to the level type to store this information. The toString function
basically maps over a two dimensional list of coordinates, pulls the
tiles at this position out of the level and renders the top most to a
char.